return public exceptions to the ignite-api module
