+++ modules/client/src/main/java/org/apache/ignite/client/BasicAuthenticator.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.client;
 
 
-import org.apache.ignite.security.authentication.AuthenticationType;
+import org.apache.ignite.security.AuthenticationType;
 
 /**
  * Basic authenticator with username and password.
