+++ modules/security/src/main/java/org/apache/ignite/internal/security/authentication/AuthenticatorFactory.java
@@ -19,7 +19,7 @@ package org.apache.ignite.internal.security.authentication;
 
 import org.apache.ignite.internal.security.authentication.configuration.AuthenticationProviderView;
 import org.apache.ignite.internal.security.authentication.configuration.BasicAuthenticationProviderView;
-import org.apache.ignite.security.authentication.AuthenticationType;
+import org.apache.ignite.security.AuthenticationType;
 
 /** Factory for {@link Authenticator}. */
 class AuthenticatorFactory {
